debounced delete fields but this affects all delete buttons
